Move MapData to org.apache.spark.sql.catalyst.data.
