+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/SpecializedGetters.java
@@ -21,7 +21,7 @@ import org.apache.spark.sql.catalyst.data.InternalRow;
 import org.apache.spark.sql.catalyst.data.ArrayData;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.Decimal;
-import org.apache.spark.sql.catalyst.util.MapData;
+import org.apache.spark.sql.catalyst.data.MapData;
 import org.apache.spark.unsafe.types.CalendarInterval;
 import org.apache.spark.unsafe.types.UTF8String;
